Add league/oauth-server and start to split route providers by group.
