+++ composer.json
@@ -31,7 +31,8 @@
         "kzykhys/ciconia": "~1.0",
         "aptoma/twig-markdown": "0.2.*",
         "igorw/config-service-provider": "~1.2",
-        "symfony/yaml": "~2.5"
+        "symfony/yaml": "~2.5",
+        "league/oauth2-server": "^4.1"
     },
     "autoload": {
         "psr-0": {"OpenCFP": "classes/"}
